requirements.txt
aiogram==3.4.1
aiocryptopay==0.3.5
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
